port: Add extra handling for FAULTY port
When a currently active SLAVE port recognizes a temporary fault (and
enters FAULTY state, while port remains available), the redundancy port
does not enter SLAVE state but remains PASSIVE.
Instead, force the PASSIVE port to step in as SLAVE for the FAULTY port.
When the previously FAULTY port is recovering, it does not need to go
through UNCALIBRATED state, but it can instantly replace the redundant
port as SLAVE again.

Signed-off-by: Stephan Wurm <stephan.wurm@a-eberle.de>