IEC62439-3: Add doubly-attached clock support
# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

This patch series adds support for doubly-attached clocks as described
in IEC 62439-3:2016 Appendix A.
It is used for clocks with redundant network connection via HSR or PRP.
The specification also defines additional profile ids for single- and
doubly-attached clocks.

Both interfaces of a redundant connection need to be handled
individually, although they are connected to the same grandmaster.
The synchronized port in SLAVE state is then backed-up by the redundant
port, which remains in PASSIVE state, unless the former port enters
FAULTY state. The PASSIVE redundant port can then step in immediately.

This improves the PTP availability in high-assurance networks, e.g. for
power automation and industrial networks.

In order to use the new doubly-attached clock feature, the
`dataset_comparison` configuration option has to be set to `iec62439-3`
and a redundancy pair needs to be defined by adding `paired_interface`
option with the respective other interface to the configuration.

The patch series also adds a default configuration for clocks according
to the Power Utility Profile as defined in IEC 61850-9-3:2016, relying
on the doubly-attached clock feature.
See also https://en.wikipedia.org/wiki/IEC/IEEE_61850-9-3 for a brief
introduction.

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.
To: linuxptp-devel@lists.nwtime.org
Signed-off-by: Stephan Wurm <stephan.wurm@a-eberle.de>

---
Changes in v3:
- Rebased on v4.4

Changes in v2:
- port.h: Removed unrelated whitespace changes
- port.c: Always translate internal port state PS_PASSIVE_SLAVE to
  PS_PASSIVE in management tlv
- pmc.c: Restored original interpretation of port state
- ptp4l.c: Improved variable naming
- ptp4l.8: Fixed naming of configuration parameter
- configs: Split into server and client-only use cases

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 3,
    "change-id": "20230317-iec62439-3-251804d04825",
    "base-branch": "feature/iec61850-9-3",
    "prefixes": [],
    "history": {
      "v1": [
        "20230317-iec62439-3-v1-0-eea0a329a7fc@a-eberle.de"
      ],
      "v2": [
        "20230317-iec62439-3-v2-0-b551a3a89314@a-eberle.de"
      ]
    }
  }
}