Fix PRP trailer detection
# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

The PRP standard is defined by IEC62493-3 and allows for redundant
connections, with automatic removal of duplicate frames and trailer
data from the logical PRP interface. In contrast, PTP has to handle
the redundant lower layer interfaces individually, hence it has to
detect the PRP trailer and remove it from the message on its own.
This PRP trailer handling was introduced with commit
02c7ab8436ced2e92524bb7b196189c8d381f883.

During tests with different PRP setups, we now found some issues with
the PRP trailer detection.

In one setup with our system connected via PRP to a redundancy device
(Siemens RedBox), the communication included all relevant messages, but
ptp4l reported

   port 1 (eth2): bad message
   port 1 (eth2): bad message
   port 1 (eth2): delay timeout

As a consequence, the port state was alternating between UNCALIBRATED
and FAULTY forever.

We found the PRP trailer not being removed from Sync and Follow_Up
messages, that had a length of exactly 66 octets, including 14 octets
ethernet header, 44 octets PTP message, 2 octets of padding and 6 octets
trailer.

Looking at the original implementation, we found the PRP_MIN_PACKET_LEN
to be configured to 70 octets, which would be correct for a whole
ethernet frame. But under Linux, a packet socket returns only ethernet
header and payload data, but not the frame check sequence (FCS) that is
part of the frame size calculation.

In another setup, we were using a MOXA RedBox with Power Profile
(IEEE C37.238-2011) that includes a mandatory VLAN header. When using
ptp4l on top of the physical ethernet interfaces (eth2, eth3), the ports
were also stuck to UNCALIBRATED, even after shortening the
PTP_MIN_PACKET_LEN to 66 octets.
We found that Linux also removes VLAN headers (802.1Q and 802.1ad) from
frames received from packet sockets, unless explicitly requesting all
additional headers. But this would only work when operating on the
physical interfaces. When using the logical VLAN interfaces (here
eth2.1 and eth3.1), the VLAN headers are stripped again and the
padding calculation in the original PRP trailer detection would fail
again for Sync and Follow_Up messages.

As a consequence, we removed both the PRP_MIN_PACKET_LEN and padding
calculation and are using always the last six octets of a message to
detect a PRP trailer, as long as it would fit between the end of a
PTP message and the end of a frame (before the FCS).

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.
To: linuxptp-devel@lists.sourceforge.net
Cc: Magnus Armholt <magnus.armholt@gmail.com>
Fixes: 02c7ab8436ced2e92524bb7b196189c8d381f883
Signed-off-by: Stephan Wurm <stephan.wurm@a-eberle.de>
Reviewed-by: Johannes Eigner <johannes.eigner@a-eberle.de>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20230623-fix-prp-trailer-detection-v1-1-6b0cc25b2f60@a-eberle.de



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20230623-fix-prp-trailer-detection-f8d0b88a7128",
    "base-branch": "master",
    "prefixes": [],
    "history": {
      "v1": [
        "20230623-fix-prp-trailer-detection-v1-1-6b0cc25b2f60@a-eberle.de"
      ]
    }
  }
}